Update patch set 2
Patch Set 2:

This is due to https://git.opendaylight.org/gerrit/#/c/36117/ which was merged yesterday. It removed throwing the UnknownMessageException in ShardTransaction. I don't know how that patch passed the test. We could just remove the test as it doesn't add much value but ignoring  is fine too.

Patch-set: 2