Update patch set 4
Patch Set 4:

With the config module check gone, there's still a timing issue with the ToasterTest in that the ToasterService RPC (i.e. OpendaylightToaster) isn't registered yet. This is a non-OSGi service dependency where the toaster-provider has to be started first. I'm going to work on extending the "rpcService" blueprint extension to wait for the underlying DOM RPC service to be available.

Patch-set: 4
Label: Verified=0