Update patch set 4
Patch Set 4:

> What is your "pro"?

I have stumbled [0] upon one. ${project.*} is evaluated at a called, so if the caller inherits from a parent pom which also uses ${project.*}, it will resolve to values wring from parent point of view.

So there is a risk of the same property name to be used more than once in inheritance chain. But a naming convention could recude the risk of that happening.

[0] https://git.opendaylight.org/gerrit/#/c/40058/1..2/karaf/karaf-parent/pom.xml

Patch-set: 4