Update patch set 6
Patch Set 6:

Ouch, looks like I did have Posted a comment on a wrong Change. My -1 comment on patch set 13 of https://git.opendaylight.org/gerrit/39400 was meant to be posted here just before patch set 5 upload.
Here is a copy:

The aries.proxy.impl version override introduced in https://git.opendaylight.org/gerrit/36801
does not result in the new version being available in the resulting Karaf installation, probably because karaf-plugin only searches for dependencies from features, so it does see package the direct dependency inherited from parent pom.

What is worse, the packaging process (for some reason) picks up the 1.0.5 version if it is already present in local Maven repository, which is what happened in the verify job (as /tmp/r was not cleared up, and the slave was building distribution without this patch previously).

The two previous paragraphs describe two bugs, the first one has to be fixed before this can be merged.

Patch-set: 6