Bug 6030 - "karaf" script invokes /bin/sh but requires /bin/bash functions
The bin/karaf script uses the "local" command which is a shell builtin of bash
and similar shells, but is not required for POSIX-compliance in sh.

On most flavors of linux, this resolves to bash or dash which probably runs in
a restricted environment after checking to see that its $0 is sh. But on Solaris's
/bin/sh is actually ksh93 for backwards compatibility.

Change-Id: Ifdd0e9fe798e881df01e7d2b586bc7e7142c0730
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>
(cherry picked from commit ddeacdd4072121c423f478316c5cc00192f9312b)