Update patch set 46
Patch Set 46:

(2 comments)

The TransactionIdentifier toString() is outputted extensively for debugging and is important for tracking txns. The current toString() is going to be too long for the logging (and gripping). I'd like to keep it compact and similar to the prior format, e.g. member-1-txn-5157-1465012553130. The last part was the timestamp which you've gotten rid of.

Patch-set: 46