Update patch set 25
Patch Set 25:

The exceptions are named the same and identical to the ones in the cds-access. It's nice to avoid related classes with the same name. Can we just use the ones in cds-access, use a short instead of an enum for closest version? They're just exceptions so probably don't need to be that precise.

Patch-set: 25
Label: Verified=0