Stop building test-jar for sal-dummy-distributed-datastore
Newer versions of maven-jar-plugin no longer support multiple outputs
with the same classifier, so the dual "jar" and "test-jar" goals here
cause an error. Since test-jar ends up being empty, and it isn't used
anywhere else, this patch simply drops the maven-jar-plugin
configuration so it only builds a standard JAR (and the shaded
all-in-one).

Change-Id: Iadc3def3df2f4ff1e9ac647c538b9fe2c4001ca4
Signed-off-by: Stephen Kitt <skitt@redhat.com>