Update patch set 1
Patch Set 1:

I know it's useful for debugging. The problem we ran into in Be was that a toString was changed that outputted the whole data tree and it was included in an error message (twice actually, before and after). This caused OOM in integration tests. I've also seen the error on a production system although it didn't run OOM. To fix it, we kept the toString but changed the error message to not output the toString. So it's OK as long as it's outputted only in debug messages but of course it's hard to tell all the call sites. We'll see how it goes - hopefully these changes won't cause an issue.

Patch-set: 1