Add "static-reference" blueprint extension
Added a blueprint extension, "static-reference", that obtains an OSGi
service and returns the actual instance. This differs from the standard
"reference" element that returns a dynamic proxy whose underlying service
instance can come and go.

I modified the binding-broker.xml to use this new extension. The static
reference is needed for the BindingToNormalizedNodeCodec in order to support
restarts of the sal-binding-broker blueprint container as the
BindingToNormalizedNodeCodec instance is created statically. Using the
standard "reference" caused errors after the container restart b/c the
ClassLoadingStrategy service proxy held onto by the BindingToNormalizedNodeCodec
was invalid. Both the ClassLoadingStrategy and SchemaService instances exist for
the life of the karaf container.

Change-Id: I4cbcc7e2b5a85b0a22e50e12f3946d29bfb36c7d
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>