Remove global BindingToNormalizedNodeCodec instance
The BindingToNormalizedNodeCodec was made a global static instance for
backwards compatibility for CSS users that inject the binding-dom-mapping-service
identity which defines the provided service as the concrete
BindingToNormalizedNodeCodec class instead of an interface. Therefore
the global static instance was created via blueprint and advertised via
its interfaces and was obtained via the static reference by the
RuntimeMappingModule for use by CSS users. The RuntimeMappingModule must
return an instance of BindingToNormalizedNodeCodec in order to provide
the binding-dom-mapping-service and obtaining and obtaining the
blueprint advertised OSGi service via its interfaces and casting to
BindingToNormalizedNodeCodec failed b/c Aries creates a service proxy
which loses the fact that it's a BindingToNormalizedNodeCodec instance.

However the global BindingToNormalizedNodeCodec is not clean and
problematic for supporting blueprint container restarts. Aries supports
concrete class proxies so I added a service export for the
BindingToNormalizedNodeCodec class as well as its interfaces. The
RuntimeMappingModule now obtains the BindingToNormalizedNodeCodec
instance from the OSGi registry as the other blueprint-brige CSS modules
do. This eliminates the need for the global instance.

Change-Id: I969ad5470967a81b37078393701c69d1898086cd
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>