Update patch set 1
Patch Set 1:

The BindingAwareBroker is actually provided by CSS. I did not provide it via blueprint b/c it isn't necessary, it's an artifact of CSS. But the CSS does also advertise it as an OSGi service so that's why it's available in the service registry. I thought the IT tests were working - was there a failure that prompted this change? Do we need the base class to provide the BindingAwareBroker or can we just remove it?

Patch-set: 1