Update patch set 1
Patch Set 1: Code-Review-2

It looks like this is to convert CDS EOS to the new APIs. I already have patches to do this but Robert wanted to hold off on it. It looks like you've just copied the previous code. What I did was to convert  CDS EOS to implement the new APIs and create an adapter layer between the "old" and new APIs. If we intend to convert CDS at this point, we can use my patches. I had it all working between "old" and new APIs. If you want to take over the patches that would be fine.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2