BUG-5280: rework SequencedQueue
Upper layers require that we retain sequence of events
across multiple request targets sharing the same shard
(i.e. cookie).

This is important, for example, to ensure the backend
see free-standing transactions originating from one
thread in correct sequence -- otherwise they could get
reordered, leading to developer confusion.

While this could be achieved with some magic in upper
layers (like it's done in TransactionProxy & co.),
such a solution is overly complicated as it involves
chaining otherwise unrelated entities via
CompletionStages (or similar).

The same effect can be achieved in ClientActorContext
by having a per-cookie request queue, not per-target
as it is currently implemented and having that queue
dispatch requests in order.

In order to do that, though, sequence numbers need to
be assigned on a per-cookie basis -- while being really
transport-specific. That requires a coordinated rework
of how requests are constructed and enqueued, which
will be done in a follow-up patch.

This patch makes SequencedQueue smart enough to handle
multiple targets and keeps one queue per cookie.

Change-Id: I11537603b95ef78e2ecbe71080712f37fb9d3e6e
Signed-off-by: Robert Varga <rovarga@cisco.com>