Update patch set 4
Patch Set 4:

(1 comment)

Is that the piece of code adding the capability to add default-config from xml file rather than from blueprint and yang?
If so, it will be possible to add this file once the distribution is built, or by adding the file in the karaf bundle and override the deployment of the attached source to ensure the file will be present in the compiled distribution. Isn't it?

Patch-set: 4