Change cancel() behavior of PingPongTransactionChain
The cancel() method was implemented as UnsupportedOperation, which
may get the transaction chain into trouble that an chain with unsubmitted
transaction can never be properly disposed.  In order to close a chain, all
open transactions must be submitted so that the chain can be unlocked.
However, this is against our purpose that we want to cancel the transactions
and may cause data inconsistence.

This patch is to fix the problem. It marks the chain as "failed" make it unable
to create new transaction. But the chain can be properly closed when failure happens.

Change-Id: Ie38a1f1ae33153c6d6dcc9c816001b7efe7b437e
Signed-off-by: Kevin Wang <kevixw@gmail.com>