Add shared PeerInfoCache
Moved the peerInfoMap from RaftActorContextImpl to a new
WriteablePeerInfoCache private inner class. It's base class, PeerInfoCache,
is a public interface providing a thread-safe, read-only view. The
PeerInfoCache can be safely accessed outside of the RaftActor and will be
used in a subsequent patch for access by a separate actor. The
WriteablePeerInfoCache utilizes copy-on-write semantics when updating the
internal peerInfoMap as updates are infrequent. Updates are not synchronized
as write access is only allowed via the RaftActorContext interface which is
owned by the RaftActor.

Change-Id: I14fabbfd2a796b50a733aa2e89605c21f446c8b0
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>