Add KeepAlive actor
Added a new KeepAliveActor that is created by the Leader behavior to
periodically send KeepAlive messages to all followers. This is done
to alleviate unwanted follower election timeouts that can occur if the
leader is too busy for some reason and AppendEntries heartbeats are
delayed. Since AppendEntries heartbeats are sent at the heartbeat
interval (500 ms default), normally they will prevent election timeouts
so we don't need the KeepAlive messages sent that frequently so the time
interval in between KeepAlive messages is one-fifth the election timeout
(or 4 times the heartbeat interval, by default 2 s).

The KeepAliveActor utilizes the shared read-only PeerInfoCache to obtain
the peer follower addresses.

Change-Id: Ic58abefc6f6138a336a236a6e99c196b6a54fd50
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>