Update patch set 2
Patch Set 2:

Why not change RaftRPC to extend Externalizable ? I guess this would avoid having proxy, but for sure it will require much more work as all the messages will have to be adapt.

Is there any reason why you selected AppendEntries and AppendEntriesReply? Are those ones the biggest? 

Also, the implies that devs will have to maintain those methods, but AFAIK so far only Robert and you (Ryan?) are the "devs" so that shouldn't be a problem.

Patch-set: 2