Update patch set 1
Patch Set 1:

> Why do we want to remove src/main/yang as a source folder in
 > eclipse? I like it as a source folder.

Tom, this goes together with https://git.opendaylight.org/gerrit/#/c/42585/.  Extensive analysis and reason why this needs to be change can be found in https://bugs.opendaylight.org/show_bug.cgi?id=6252.. basically, *.yang resources are expected to be on the Classpath on META-INF/yang by much of our code, and not the root, so having src/main/yang (with *.yang directly in it) as a source folder is WRONG.  I'm the one how initially introduced it, but now realize this was a mistake, and is causing problems such as the exception shown in Bug 6252 when writing JUnit tests which read *.yang files from the classpath. PS, FYI: I'm off on vacation from tomorrow until end of next week, thus next reply only after I return.

Patch-set: 1