Update patch set 2
Patch Set 2: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/controller-validate-autorelease-boron/1101/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

- groupId is correct and no typos
- artifactId is correct and no typos
- version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/controller-validate-autorelease-boron/1101/ )

https://jenkins.opendaylight.org/releng/job/controller-distribution-check-boron/2945/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

- groupId is correct and no typos
- artifactId is correct and no typos
- version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/controller-distribution-check-boron/2945/ )

https://jenkins.opendaylight.org/releng/job/controller-verify-boron-mvn33-openjdk8/6/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

- groupId is correct and no typos
- artifactId is correct and no typos
- version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/controller-verify-boron-mvn33-openjdk8/6/ )

Patch-set: 2