Update patch set 3
Patch Set 3:

> I think this is affected by the same thing as the other patch.
 > Let's hold it until boron is branched.

yeah but this is a lot less likely to be an issue. If a shard snapshotted in He, most likely it has also snapshotted in Li and Be. But it would be safer to keep it around for another release.

Patch-set: 3
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0