Update patch set 26
Patch Set 26:

I was thinking about that -- and I think this is something that should be done as a simple counter on a per-client basis, so we preserve fairness.

I am not sure if we have sufficiently-defined notion of a client connection in ShardCommitCoordinator (i.e. old messages). In the new messages we have an explicit 'ConnectClient' message and can advertise the limit back to the client -- and we will have a LeaderFrontendState (https://git.opendaylight.org/gerrit/41032) to track it. We can even add explicit XON/XOFF notifications to provide pushback.

Patch-set: 26