Update patch set 26
Patch Set 26:

> I was thinking about that -- and I think this is something that
 > should be done as a simple counter on a per-client basis, so we
 > preserve fairness.
 > 
 > I am not sure if we have sufficiently-defined notion of a client
 > connection in ShardCommitCoordinator (i.e. old messages). In the
 > new messages we have an explicit 'ConnectClient' message and can
 > advertise the limit back to the client -- and we will have a
 > LeaderFrontendState (https://git.opendaylight.org/gerrit/41032) to
 > track it. We can even add explicit XON/XOFF notifications to
 > provide pushback.

There isn't a way to do it per-client in SCC. I don't think we need to preserve it right now as I have yet to see the queue limit reached (the rate limiting does a good job of keeping the queue size low). So I'm inclined to bag it at least for now - it would be throw away in ShardDataTree until your frontend stuff is complete. 

Either way I'll merge this patch. We can always follow-up if we think we want to preserve it in Boron.

Patch-set: 26