WIP: BUG-5280: separate request sequence and transmit sequence
Clean up the confusion in sequence numbers. There are actually
two sequences.

The first one is the logical request sequence, which indicates
the order of requests in which they should be applied to the target
entity. It is assigned by logic emitting the request.

The second one is the transmit sequence within a connection
to the backend, as initiated by ConnectClientRequest. It is assigned
by SequencedQueue as it is transmitting requests.

Change-Id: I61663073bf6632c1ed8c036dee37f1ac39cf7794
Signed-off-by: Robert Varga <rovarga@cisco.com>