Prevent trigger  of unnecessary leader elections
The ElectionTimeout contains a count to ensure that we process the most
recent message. Due to slow processing of AppendEntries, it might be
possible that election times out on the follower and a new one is
triggered. To prevent a pre-mature election a follower only processes
the one which was scheduled the latest.

TODO: Unit tests need to be fixed.

Change-Id: I2356f9e1eb36700294a8c83d96f7536cbef1f7e6
Signed-off-by: Faiz Ilahi Kothari <faiz.ilahi.k.kothari@ericsson.com>