Update patch set 1
Patch Set 1: Code-Review-1

I'm creating this Gerrit to "park" working (!) code which I originally wrote in the context of https://git.opendaylight.org/gerrit/#/c/42109/, but currently no longer need there.

-1 because I do not require this at this point. If this is of (possible) interest to anyone else, feel free to merge.

If I do not see any need in the next 2 weeks, I'll close it as Abandon, to "park" this, either forever, or until I have a need for it, and then revive it.

Patch-set: 1
Label: Code-Review=-1