Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

This is not related to clustering and does not belong in sal-distributed-datastote. I would suggest sal-dom-broker. 

Also I'd prefer not to prefix the classes with Mdsal - it doesn't provide useful meaning as everything in project is mdsal-related.  I would suggest using the Adapter suffix.

Patch-set: 3
Label: Code-Review=-1