Update patch set 1
Patch Set 1:

This message is benign. In fact it's better that it doesn't create a proxy. There's really no need for it - the only reason it does is due to the presence of the Apache quiesce stuff that adds a participant that keeps counters of method calls (for what use? I don't know). If the quiesce bundle wasn't present it wouldn't add the interceptor proxy. Removing the quiesce stuff from our distro is an option but not sure if there would be any negative impact.

Another option is to write our own <service> extension to avoid the proxy overhead. I did this with service references with <static-reference>.

The simplest solution in this case is to remove final.

Patch-set: 1
Label: Code-Review=0