Update patch set 3
Patch Set 3:

(25 comments)

Can you explain a bit how CreatePrefixedShard will be used? Will it be invoked via the ClusterAdminService by a user?

A while back Moiz and I had envisioned an RPC on the ClusterAdminService to add a shard configuration. The message to the ShardManager would add the config and create a local shard with no peers. The shard config would be persisted in the ShardManagerSnapshot. If the user specified other member replicas, then the ClusterAdminService would send AddShardReplica to those members (and report any failures). In onAddShardReplica, if no shard config existed, it would query the other members for the shard config and, if found, would proceed with AddServer. If the user wanted to later add another member replica or if AddShardReplica initially failed, the user would issue the add-shard-replica RPC.

Patch-set: 3